requirements.txt
# Steam Dataset 2025 - Python Dependencies
# Pinned versions from working development environment
# Last updated: 2025-09-02

# Core Data Processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Database Connectivity
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# HTTP Requests and API Integration
requests==2.31.0
urllib3==2.1.0

# Environment and Configuration Management
python-dotenv==1.0.0

# JSON and Data Validation
jsonschema==4.20.0

# Command Line Interface
argparse>=1.4.0  # Built-in, version varies by Python
tqdm==4.66.1

# Machine Learning and Vector Operations
scikit-learn==1.3.2
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.36.2

# Vector Database Support
pgvector==0.2.4

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Jupyter and Interactive Analysis
jupyter==1.0.0
ipython==8.18.1
notebook==7.0.6

# Text Processing and NLP
nltk==3.8.1
spacy==3.7.2

# Statistical Analysis
statsmodels==0.14.1

# Development and Testing
pytest==7.4.3
black==23.12.0
flake8==6.1.0
mypy==1.8.0

# Progress Bars and Utilities
rich==13.7.0
click==8.1.7

# Date and Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# File Format Support
openpyxl==3.1.2
xlsxwriter==3.1.9

# Memory Profiling (Optional)
memory-profiler==0.61.0

# API Rate Limiting
ratelimit==2.2.1

# Graph Database Support (Future)
neo4j==5.15.0

# Additional Scientific Computing
umap-learn==0.5.4
gensim==4.3.2

# Documentation Generation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0